Remove id from /wards/[id]/schedule-success
